add animation to carousel points

See merge request boran/dongu_mobile!43

+++ lib/presentation/screens/filter_view/components/custom_sliderbar_textfield.dart
@@ -10,16 +10,14 @@ class CustomSliderBarAndTextField extends StatefulWidget {
   CustomSliderBarAndTextField({Key? key}) : super(key: key);
 
   @override
-  _CustomSliderBarAndTextFieldState createState() =>
-      _CustomSliderBarAndTextFieldState();
+  _CustomSliderBarAndTextFieldState createState() => _CustomSliderBarAndTextFieldState();
 }
 
-class _CustomSliderBarAndTextFieldState
-    extends State<CustomSliderBarAndTextField> {
-  double _starValue = 10;
-  double _endValue = 50;
-  double minValue = 10.0;
-  double maxValue = 500.0;
+class _CustomSliderBarAndTextFieldState extends State<CustomSliderBarAndTextField> {
+  int _starValue = 10;
+  int _endValue = 50;
+  int minValue = 10;
+  int maxValue = 500;
 
   final startController = TextEditingController();
   final endController = TextEditingController();
@@ -32,28 +30,26 @@ class _CustomSliderBarAndTextFieldState
   }
 
   _setStartValue() {
-    if (double.parse(startController.text).roundToDouble() <=
-            double.parse(endController.text).roundToDouble() &&
+    if (double.parse(startController.text).roundToDouble() <= double.parse(endController.text).roundToDouble() &&
         double.parse(startController.text).roundToDouble() >= minValue &&
         double.parse(endController.text).roundToDouble() >= minValue &&
         double.parse(startController.text).roundToDouble() <= maxValue &&
         double.parse(endController.text).roundToDouble() <= maxValue) {
       setState(() {
-        _starValue = double.parse(startController.text).roundToDouble();
+        _starValue = double.parse(startController.text).roundToDouble().toInt();
       });
     }
     print("Second text field: ${startController.text}");
   }
 
   _setEndValue() {
-    if (double.parse(startController.text).roundToDouble() <=
-            double.parse(endController.text).roundToDouble() &&
+    if (double.parse(startController.text).roundToDouble() <= double.parse(endController.text).roundToDouble() &&
         double.parse(startController.text).roundToDouble() >= minValue &&
         double.parse(endController.text).roundToDouble() >= minValue &&
         double.parse(startController.text).roundToDouble() <= maxValue &&
         double.parse(endController.text).roundToDouble() <= maxValue) {
       setState(() {
-        _endValue = double.parse(endController.text).roundToDouble();
+        _endValue = double.parse(endController.text).roundToDouble().toInt();
       });
     }
     print("Second text field: ${endController.text}");
@@ -148,9 +144,9 @@ class _CustomSliderBarAndTextFieldState
                   child: TextField(
                     keyboardType: TextInputType.number,
                     decoration: InputDecoration(
-                        border: InputBorder.none, hintText: LocaleKeys.filters_package_price_item3.locale,
-                        hintStyle: AppTextStyles.subTitleStyle
-                            .copyWith(fontWeight: FontWeight.bold)),
+                        border: InputBorder.none,
+                        hintText: LocaleKeys.filters_package_price_item3.locale,
+                        hintStyle: AppTextStyles.subTitleStyle.copyWith(fontWeight: FontWeight.bold)),
                     controller: endController,
                     cursorColor: AppColors.textColor,
                   ),
@@ -207,8 +203,9 @@ class _CustomSliderBarAndTextFieldState
                   child: TextField(
                     keyboardType: TextInputType.number,
                     decoration: InputDecoration(
-                        border: InputBorder.none, hintText: LocaleKeys.filters_package_price_item3.locale,
-                         hintStyle: AppTextStyles.subTitleStyle.copyWith(fontWeight: FontWeight.bold)),
+                        border: InputBorder.none,
+                        hintText: LocaleKeys.filters_package_price_item3.locale,
+                        hintStyle: AppTextStyles.subTitleStyle.copyWith(fontWeight: FontWeight.bold)),
                     controller: startController,
                     cursorColor: AppColors.textColor,
                   ),
@@ -224,21 +221,19 @@ class _CustomSliderBarAndTextFieldState
 
   SliderTheme rangeSliderBar() {
     return SliderTheme(
-      data: SliderThemeData(
-        trackHeight: 7
-      ),
+      data: SliderThemeData(trackHeight: 7),
       child: RangeSlider(
-        values: RangeValues(_starValue, _endValue),
-        min: minValue,
-        max: maxValue,
+        values: RangeValues(_starValue.toDouble(), _endValue.toDouble()),
+        min: minValue.toDouble(),
+        max: maxValue.toDouble(),
         inactiveColor: AppColors.sliderColor,
         activeColor: AppColors.greenColor,
         onChanged: (values) {
           setState(() {
-            _starValue = values.start.roundToDouble();
-            _endValue = values.end.roundToDouble();
-            startController.text = values.start.roundToDouble().toString();
-            endController.text = values.end.roundToDouble().toString();
+            _starValue = values.start.roundToDouble().toInt();
+            _endValue = values.end.roundToDouble().toInt();
+            startController.text = values.start.roundToDouble().toInt().toString();
+            endController.text = values.end.roundToDouble().toInt().toString();
           });
         },
       ),
